Optionally remove unreachable commits in rake task.
When the repository grows with time a 'git fsck' takes
too long to be run on every fetch of the repository.
This might not be an issue when running periodically
with cron, but when using webhooks the request times
out for big repositories.

To workaround this problem, the user might leave the
'Remove unreachable on fetch' global setting unchecked
to skip the check while doing the fetch. Instead a
separate rake task can be used to remove any unreachable
commits periodically:

rake redmine:plugins:git_mirror:remove_unreachable_commits \
	RAILS_ENV="production"

Signed-off-by: Florin Hillebrand <florin.hillebrand@tuwien.ac.at>